requirements.txt
streamlit
FinanceDataReader
google-generativeai